+++ MecabCocoa/dictionary.cpp
@@ -244,7 +244,7 @@ bool Dictionary::compile(const Param &param,
       token->rcAttr = rid;
       token->posid  = pid;
       token->wcost = cost;
-      token->feature = offset;
+      token->feature = (unsigned int) offset;
       token->compound = 0;
       dic.push_back(std::make_pair<std::string, Token*>(w, token));
 
@@ -310,11 +310,11 @@ bool Dictionary::compile(const Param &param,
   }
 
   unsigned int dummy = 0;
-  unsigned int lsize = matrix.left_size();
-  unsigned int rsize = matrix.right_size();
-  unsigned int dsize = da.unit_size() * da.size();
-  unsigned int tsize = tbuf.size();
-  unsigned int fsize = fbuf.size();
+  unsigned int lsize = (unsigned int) matrix.left_size();
+  unsigned int rsize = (unsigned int) matrix.right_size();
+  unsigned int dsize = (unsigned int) da.unit_size() * (unsigned int) da.size();
+  unsigned int tsize = (unsigned int) tbuf.size();
+  unsigned int fsize = (unsigned int) fbuf.size();
 
   unsigned int version = DIC_VERSION;
   char charset[32];
