+++ MecabCocoa/tokenizer.cpp
@@ -149,10 +149,10 @@ bool TokenizerImpl<N, P>::open(const Param &param) {
     d->next = dictionary_info_;
     d->filename = dic_[i]->filename();
     d->charset = dic_[i]->charset();
-    d->size = dic_[i]->size();
-    d->lsize = dic_[i]->lsize();
-    d->rsize = dic_[i]->rsize();
-    d->type = dic_[i]->type();
+    d->size = (unsigned int) dic_[i]->size();
+    d->lsize = (unsigned int) dic_[i]->lsize();
+    d->rsize = (unsigned int) dic_[i]->rsize();
+    d->type = (unsigned int) dic_[i]->type();
     d->version = dic_[i]->version();
     dictionary_info_ = d;
   }
